Upgrade sodium-native
Problem: sodium-native has released a new major version and it's great
to stay up-to-date with our upstream dependencies.

Solution: Upgrade it. It seems that the breaking change is "Removed
object instance apis and replaced them with init, update, final
methods", but that doesn't seem to apply to us (?) as `npm test`
continues to pass without problems.